requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0

# Agent Dependencies
pydantic-ai[gemini]==0.0.13
google-adk==0.4.0

# Observability
logfire>=0.46.0
structlog==23.2.0
arize==7.46.0
arize-otel>=0.3.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0

# OpenInference for LLM instrumentation
openinference-instrumentation-openai>=0.1.0
openinference-instrumentation-google-genai>=0.1.0

# Phoenix Evaluations
arize-phoenix[evals]>=0.8.0

# Cost Tracking
tokencost>=0.1.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
ruff==0.1.6
black==23.11.0
mypy==1.7.1
pre-commit==3.6.0

# Additional Testing
dirty-equals==0.7.1
faker==20.1.0

# CLI Dependencies
click==8.1.7
rich==13.7.0

# Utility Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# HTTP Client
aiohttp==3.9.1

# Development Tools
ipython==8.17.2
jupyter==1.0.0

# Build Tools
build==1.0.3
wheel==0.42.0

# Security
cryptography==41.0.7

# Data Validation
email-validator==2.1.0
phonenumbers==8.13.25

# Performance
uvloop==0.19.0

# Monitoring
prometheus-client==0.19.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Type Checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12